+++ src/main/java/com/mobble/mobbleserver/application/chat/room/port/provided/direct/DirectChatRoomCreatePort.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.application.chat.room.port.provided.direct;
 
-import com.mobble.mobbleserver.refactor.chat.directChatRoom.dto.request.DirectChatRoomCreateRequestDto;
-import com.mobble.mobbleserver.refactor.chat.directChatRoom.dto.response.DirectChatRoomPreviewResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.chat.room.direct.dto.request.DirectChatRoomCreateRequestDto;
+import com.mobble.mobbleserver.infrastructure.web.chat.room.direct.dto.response.DirectChatRoomPreviewResponseDto;
 
 public interface DirectChatRoomCreatePort {
