+++ src/main/java/com/mobble/mobbleserver/application/chat/room/port/provided/direct/DirectChatRoomQueryPort.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.application.chat.room.port.provided.direct;
 
-import com.mobble.mobbleserver.refactor.chat.directChatRoom.dto.response.DirectChatRoomPreviewResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.chat.room.direct.dto.response.DirectChatRoomPreviewResponseDto;
 
 import java.util.List;
