+++ src/main/java/com/mobble/mobbleserver/refactor/chat/clubChatRoom/service/ClubChatRoomService.java
@@ -5,8 +5,8 @@ import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.refactor.chat.chatMessage.dto.request.ChatMessageRequestDto;
 import com.mobble.mobbleserver.refactor.chat.chatMessage.dto.response.ChatMessageResponseDto;
 import com.mobble.mobbleserver.refactor.chat.chatMessage.service.ChatMessageService;
-import com.mobble.mobbleserver.refactor.chat.clubChatRoom.dto.request.ClubChatMessageRequestDto;
-import com.mobble.mobbleserver.refactor.chat.clubChatRoom.dto.response.ClubChatMessageResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.chat.room.club.dto.request.ClubChatMessageRequestDto;
+import com.mobble.mobbleserver.infrastructure.web.chat.room.club.dto.response.ClubChatMessageResponseDto;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
 import com.mobble.mobbleserver.refactor.clubMember.validator.ClubMemberValidator;
 import com.mobble.mobbleserver.util.DateTimeUtils;
